requirements.txt
django==5.0.2
django-environ==0.10.0
psycopg2-binary
django-redis==5.3.0
redis==4.5.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
celery==5.3.1
gunicorn==21.2.0
Pillow